azure-function/function_for_chat_da/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.26.4
pyyaml
pyarrow
pandasai==2.0.30
openpyxl
fastparquet
